+++ lib/percy/percy/encoders/ditt_encoder.hpp
@@ -5,7 +5,7 @@
 
 namespace percy
 {
-    class berkeley_encoder : public std_encoder
+    class ditt_encoder : public std_encoder
     {
         private:
 			int nr_op_vars_per_step;
@@ -26,13 +26,13 @@ namespace percy
             pabc::Vec_Int_t* vLits; // Dynamic vector of literals
 
         public:
-            berkeley_encoder(solver_wrapper& solver)
+            ditt_encoder(solver_wrapper& solver)
             {
                 vLits = pabc::Vec_IntAlloc(128);
                 set_solver(solver);
             }
 
-            ~berkeley_encoder()
+            ~ditt_encoder()
             {
                 pabc::Vec_IntFree(vLits);
             }
@@ -117,7 +117,7 @@ namespace percy
                 auto status = true;
 
                 if (spec.verbosity > 2) {
-                    printf("Creating op clauses (BERKELEY-%d)\n", spec.fanin);
+                    printf("Creating op clauses (DITT-%d)\n", spec.fanin);
                     printf("Nr. clauses = %d (PRE)\n", solver->nr_clauses());
                 }
 
@@ -166,7 +166,7 @@ namespace percy
                 auto status = true;
 
                 if (spec.verbosity > 2) {
-                    printf("Creating output clauses (BERKELEY-%d)\n", spec.fanin);
+                    printf("Creating output clauses (DITT-%d)\n", spec.fanin);
                     printf("Nr. clauses = %d (PRE)\n", solver->nr_clauses());
                 }
                 // Every output points to an operand.
@@ -240,7 +240,7 @@ namespace percy
                     nr_input_tt_vars + nr_sel_vars + nr_lex_vars;
 
                 if (spec.verbosity > 2) {
-                    printf("Creating variables (BERKELEY-%d)\n", spec.fanin);
+                    printf("Creating variables (DITT-%d)\n", spec.fanin);
                     printf("nr steps = %d\n", spec.nr_steps);
                     printf("nr_sel_vars=%d\n", nr_sel_vars);
                     printf("nr_op_vars = %d\n", nr_op_vars);
@@ -371,6 +371,9 @@ namespace percy
                     // need to ensure that this operand's truth table satisfies
                     // the specified output function.
                     for (int h = 0; h < spec.nr_nontriv; h++) {
+                        if (spec.is_dont_care(h, t + 1)) {
+                            continue;
+                        }
                         auto outbit = kitty::get_bit(
                                 spec[spec.synth_func(h)], t+1);
                         if ((spec.out_inv >> spec.synth_func(h)) & 1) {
@@ -400,7 +403,7 @@ namespace percy
             create_main_clauses(const spec& spec)
             {
                 if (spec.verbosity > 2) {
-                    printf("Creating main clauses (BERKELEY-%d)\n", spec.fanin);
+                    printf("Creating main clauses (DITT-%d)\n", spec.fanin);
                     printf("Nr. clauses = %d (PRE)\n", solver->nr_clauses());
                 }
                 auto success = true;