+++ lib/percy/percy/encoders/encoder.hpp
@@ -60,6 +60,8 @@ namespace percy
         virtual bool cegar_encode(const spec& spec, const fence& f) = 0;
         virtual bool create_tt_clauses(const spec& spec, int idx) = 0;
         virtual kitty::dynamic_truth_table& simulate(const spec& spec) = 0;
+
+        virtual void reset_sim_tts(int) { }
     };
 
     template<int FI>
@@ -124,8 +126,7 @@ namespace percy
         return status;
     }
 
-    inline void
-    fanin_init(std::vector<int>& fanins, int max_fanin_id)
+    inline void fanin_init(std::vector<int>& fanins, int max_fanin_id)
     {
         fanins[fanins.size()-1] = max_fanin_id--;
         for (int i = fanins.size() - 2; i >= 0; i--) {
@@ -133,8 +134,7 @@ namespace percy
         }
     }
 
-    inline void
-    fanin_init(std::vector<int>& fanins, int max_fanin_id, int start_idx)
+    inline void fanin_init(std::vector<int>& fanins, int max_fanin_id, int start_idx)
     {
         fanins[start_idx] = max_fanin_id--;
         for (int i = start_idx-1; i >= 0; i--) {
@@ -142,8 +142,7 @@ namespace percy
         }
     }
 
-    inline bool
-    fanin_inc(std::vector<int>& fanins, const int max_fanin_id)
+    inline bool fanin_inc(std::vector<int>& fanins, const int max_fanin_id)
     {
         for (auto i = 0u; i < fanins.size(); i++) {
             if (i < fanins.size() - 1) {
@@ -166,16 +165,14 @@ namespace percy
         return false;
     }
 
-    inline void
-    print_fanin(const std::vector<int>& fanins)
+    inline void print_fanin(const std::vector<int>& fanins)
     {
         for (auto i = 0u; i < fanins.size(); i++) {
             printf("%d ", fanins[i] + 1);
         }
     }
 
-    inline void
-    print_fanin(const int* const fanins, int nr_fanins)
+    inline void print_fanin(const int* const fanins, int nr_fanins)
     {
         for (auto i = 0; i < nr_fanins; i++) {
             printf("%d ", fanins[i] + 1);