+++ lib/percy/percy/encoders/partial_dag_encoder.hpp
@@ -27,10 +27,10 @@ namespace percy
         // so this is a constant.
         const int PD_OP_VARS_PER_STEP = 3;
 
-        static constexpr int NR_SIM_TTS = 32;
+        static const int NR_SIM_TTS = 32;
         std::vector<kitty::dynamic_truth_table> sim_tts { NR_SIM_TTS };
 
-        inline int nr_svars_for_step(
+        int nr_svars_for_step(
             const spec& spec, 
             const partial_dag& dag, 
             int i) const
@@ -52,6 +52,19 @@ namespace percy
             default: // No fanin flexibility
                 return 0;
             }
+        }
+
+        int nr_pi_fanins_for_step(const partial_dag& dag, int i) const
+        {
+            const auto& vertex = dag.get_vertex(i);
+            auto nr_pi_fanins = 0;
+            if (vertex[1] == FANIN_PI) {
+                nr_pi_fanins = 2;
+            } else if (vertex[0] == FANIN_PI) {
+                nr_pi_fanins = 1;
+            }
+
+            return nr_pi_fanins;
         }
 
         int get_sel_var(
@@ -511,7 +524,6 @@ namespace percy
             }
         }
 
-
         bool reapply_helper(
             const spec& spec,
             const partial_dag& dag,
@@ -667,88 +679,6 @@ namespace percy
             return reapply_helper(spec, dag, svars, 0, 0, 0, 0, 0, 0, 0, 0, 0);
         }
 
-        void colex_helper(
-            const spec& spec,
-            const partial_dag& dag, 
-            int i, 
-            int j, 
-            int k, 
-            int sel_var) 
-        {
-            int pLits[2];
-
-            const auto& vertex = dag.get_vertex(i + 1);
-            auto nr_pi_fanins = 0;
-            if (vertex[1] == FANIN_PI) {
-                nr_pi_fanins = 2;
-            } else if (vertex[0] == FANIN_PI) {
-                nr_pi_fanins = 1;
-            }
-            if (nr_pi_fanins == 1) {
-                const auto kp = spec.nr_in + vertex[1] - 1;
-                for (int jp = 0; jp < spec.nr_in; jp++) {
-                    if ((kp == k && jp < j) || (kp < k)) {
-                        const auto sel_varp = get_sel_var(spec, dag, i + 1, jp);
-                        int ctr = 0;
-                        if (sel_var != -1) {
-                            pLits[ctr++] = pabc::Abc_Var2Lit(sel_var, 1);
-                        }
-                        pLits[ctr++] = pabc::Abc_Var2Lit(sel_varp, 1);
-                        (void)solver->add_clause(pLits, pLits + ctr);
-                    }
-                }
-            } else if (nr_pi_fanins == 2) {
-                auto svar_ctr = 0;
-                for (int kp = 1; kp < spec.nr_in; kp++) {
-                    for (int jp = 0; jp < kp; jp++) {
-                        if ((kp == k && jp < j) || (kp < k)) {
-                            const auto sel_varp = get_sel_var(spec, dag, i + 1, svar_ctr);
-                            int ctr = 0;
-                            if (sel_var != -1) {
-                                pLits[ctr++] = pabc::Abc_Var2Lit(sel_var, 1);
-                            }
-                            pLits[ctr++] = pabc::Abc_Var2Lit(sel_varp, 1);
-                            (void)solver->add_clause(pLits, pLits + ctr);
-                        }
-                        svar_ctr++;
-                    }
-                }
-            }
-        }
-        
-        void create_colex_clauses(const spec& spec, const partial_dag& dag)
-        {
-            for (int i = 0; i < spec.nr_steps - 1; i++) {
-                const auto& vertex = dag.get_vertex(i);
-                auto nr_pi_fanins = 0;
-                if (vertex[1] == FANIN_PI) {
-                    nr_pi_fanins = 2;
-                } else if (vertex[0] == FANIN_PI) {
-                    nr_pi_fanins = 1;
-                }
-                if (nr_pi_fanins == 0) {
-                    const auto j = spec.nr_in + vertex[0] - 1;
-                    const auto k = spec.nr_in + vertex[1] - 1;
-                    colex_helper(spec, dag, i, j, k, -1);
-                } else if (nr_pi_fanins == 1) {
-                    const auto k = spec.nr_in + vertex[1] - 1;
-                    for (int j = 0; j < spec.nr_in; j++) {
-                        const auto sel_var = get_sel_var(spec, dag, i, j);
-                        colex_helper(spec, dag, i, j, k, sel_var);
-                    }
-                } else {
-                    auto svar_ctr = 0;
-                    for (int k = 1; k < spec.nr_in; k++) {
-                        for (int j = 0; j < k; j++) {
-                            const auto sel_var = get_sel_var(spec, dag, i, svar_ctr++);
-                            if (k == 1) continue;
-                            colex_helper(spec, dag, i, j, k, sel_var);
-                        }
-                    }
-                }
-            }
-        }
-
         bool create_symvar_clauses(const spec& spec, const partial_dag& dag)
         {
             for (int q = 1; q < spec.nr_in; q++) {
@@ -899,10 +829,6 @@ namespace percy
                 return false;
             }
 
-            if (spec.add_colex_clauses) {
-                create_colex_clauses(spec, dag);
-            }
-
             if (spec.add_symvar_clauses && !create_symvar_clauses(spec, dag)) {
                 return false;
             }
@@ -918,7 +844,7 @@ namespace percy
         /// is extracted in the CEGAR loop, this leads to trouble.
         /// For an example, try synthesizing the 4-input function with
         /// decimal truth table 127.
-        inline void create_cardinality_constraints(
+        void create_cardinality_constraints(
             const spec& spec, 
             const partial_dag& dag)
         {
@@ -995,10 +921,6 @@ namespace percy
                 return false;
             }
             
-            if (spec.add_colex_clauses) {
-                create_colex_clauses(spec, dag);
-            }
-
             if (spec.add_symvar_clauses && !create_symvar_clauses(spec, dag)) {
                 return false;
             }