+++ lib/percy/percy/encoders/ssv_dag_encoder.hpp
@@ -2,15 +2,13 @@
 
 #include "encoder.hpp"
 #include "../io.hpp"
-#include "../floating_dag.hpp"
 
 namespace percy
 {
-    class floating_dag_encoder
+    template<int FI>
+    class ssv_dag_encoder : public dag_encoder<FI>
     {
         protected:
-            solver_wrapper * solver;
-
             int nr_op_vars;
             int nr_sim_vars;
             int nr_op_vars_per_step;
@@ -18,65 +16,37 @@ namespace percy
             pabc::Vec_Int_t* vLits; // Dynamic vector of literals
 
         public:
-            floating_dag_encoder()
+            ssv_dag_encoder()
             {
                 vLits = pabc::Vec_IntAlloc(128);
             }
 
-            ~floating_dag_encoder()
+            ~ssv_dag_encoder()
             {
                 pabc::Vec_IntFree(vLits);
             }
 
-            void
-            set_solver(solver_wrapper * solver) 
-            {
-                this->solver = solver;
-            }
-
             int
-            get_sel_var(int sel_var)
+            get_op_var(const spec& spec, int step_idx, int var_idx)
             {
-            }
-
-            template<int FI>
-            int
-            get_op_var(const floating_dag<FI>& dag, int step_idx, int var_idx)
-            {
-                assert(step_idx < dag.get_nr_vertices());
+                assert(step_idx < spec.nr_steps);
                 assert(var_idx > 0);
                 assert(var_idx <= nr_op_vars_per_step);
 
                 return step_idx * nr_op_vars_per_step + var_idx - 1;
             }
 
-            template<int FI>
             int
-            get_sim_var(
-                    const spec& spec, 
-                    const floating_dag<FI>& dag,
-                    int step_idx, 
-                    int t)
+            get_sim_var(const spec& spec, const dag<FI>& dag, int step_idx, int t)
             {
                 assert(step_idx < dag.get_nr_vertices());
 
                 return nr_op_vars + spec.get_tt_size() * step_idx + t;
             }
 
-            template<int FI>
             void
-            create_variables(const spec& spec, const floating_dag<FI>& dag)
+            create_variables(const spec& spec, const dag<FI>& dag)
             {
-                auto nr_sel_vars = 0;
-                dag.foreach_vertex([&dag, &nr_sel_vars] (auto v, int v_idx) {
-                    dag.foreach_fanin(v, 
-                        [&nr_sel_vars] (int fanin, int fanin_idx) {
-                            if (fanin == pi_fanin) {
-                                // Can refer to any PI or previous step
-                            }
-                    });
-                });
-
                 nr_op_vars_per_step = ((1u << FI) - 1);
                 nr_op_vars = dag.get_nr_vertices() * nr_op_vars_per_step;
                 nr_sim_vars = dag.get_nr_vertices() * spec.get_tt_size();
@@ -86,14 +56,13 @@ namespace percy
                     printf("nr_sim_vars=%d\n", nr_sim_vars);
                 }
 
-                solver->set_nr_vars(nr_op_vars + nr_sim_vars);
+                this->solver->set_nr_vars(nr_op_vars + nr_sim_vars);
             }
 
-            template<int FI>
             bool
             add_simulation_clause(
                     const spec& spec,
-                    const floating_dag<FI>& dag,
+                    const dag<FI>& dag,
                     const int t, 
                     const int i, 
                     const int output, 
@@ -122,10 +91,10 @@ namespace percy
 
                 if (opvar_idx > 0) {
                     pabc::Vec_IntSetEntry(vLits, ctr++, pabc::Abc_Var2Lit(
-                                get_op_var(dag, i, opvar_idx), 1 - output));
+                                get_op_var(spec, i, opvar_idx), 1 - output));
                 }
 
-                auto status =  solver->add_clause(
+                auto status =  this->solver->add_clause(
                         pabc::Vec_IntArray(vLits), 
                         pabc::Vec_IntArray(vLits) + ctr); 
 
@@ -154,14 +123,10 @@ namespace percy
                 return status;
             }
 
-            template<int FI>
             bool
-            create_tt_clauses(
-                    const spec& spec, 
-                    const floating_dag<FI>& dag, 
-                    int t)
+            create_tt_clauses(const spec& spec, const dag<FI>& dag, int t)
             {
-                std::vector<int> fanins(FI);
+                int fanins[FI];
                 std::vector<int> fanin_asgn(FI);
 
                 for (int i = 0; i < dag.get_nr_vertices(); i++) {
@@ -215,12 +180,12 @@ namespace percy
                         }
                         pLits[0] = pabc::Abc_Var2Lit(get_sim_var(spec, dag, i,
                                     t), 1 - outbit);
-                        if (!solver->add_clause(pLits, pLits + 1)) {
+                        if (!this->solver->add_clause(pLits, pLits + 1)) {
                             return false;
                         }
 
                         if (spec.verbosity > 1) {
-                            printf("bit %d=%d", t+2, outbit);
+                            printf("bit %d=%lu", t+2, outbit);
                             printf("\tvar=%d\n", get_sim_var(spec, dag, i, t));
                         }
                     }
@@ -229,9 +194,8 @@ namespace percy
                 return true;
             }
 
-            template<int FI>
             bool
-            create_main_clauses(const spec& spec, const floating_dag<FI>& dag)
+            create_main_clauses(const spec& spec, const dag<FI>& dag)
             {
                 for (int t = 0; t < spec.get_tt_size(); t++) {
                     if (!create_tt_clauses(spec, dag, t)) {
@@ -241,9 +205,9 @@ namespace percy
                 return true;
             }
 
-            template<int FI>
+			/// Encodes specifciation for use in standard synthesis flow.
             bool 
-            encode(const spec& spec, const floating_dag<FI>& dag)
+            encode(const spec& spec, const dag<FI>& dag)
             {
                 bool success = true;
                 create_variables(spec, dag);
@@ -252,9 +216,9 @@ namespace percy
                 return success;
             }
 
-            template<int FI>
+			/// Encodes specifciation for use in CEGAR based synthesis flow.
             bool 
-            cegar_encode(const spec& spec, const floating_dag<FI>& dag)
+            cegar_encode(const spec& spec, const dag<FI>& dag)
             {
                 create_variables(spec, dag);
                 for (int i = 0; i < spec.nr_rand_tt_assigns; i++) {
@@ -266,22 +230,21 @@ namespace percy
                 return true;
             }
 
-            template<int FI>
+            /// Extracts chain from encoded CNF solution.
             void 
             extract_chain(
                     const spec& spec, 
-                    const floating_dag<FI>& dag, 
+                    const dag<2>& dag, 
                     chain& chain)
             {
-                assert(chain.get_fanin() == FI);
-                int op_inputs[FI];
+                std::vector<int> op_inputs(spec.fanin);
 
-                chain.reset(spec.get_nr_in(), 1, dag.get_nr_vertices());
+                chain.reset(spec.get_nr_in(), 1, spec.nr_steps, 2);
 
-                for (int i = 0; i < dag.get_nr_vertices(); i++) {
-                    kitty::static_truth_table<FI> op;
+                for (int i = 0; i < spec.nr_steps; i++) {
+                    kitty::dynamic_truth_table op(2);
                     for (int j = 1; j <= nr_op_vars_per_step; j++) {
-                        if (solver->var_value(get_op_var(dag, i, j))) {
+                        if (this->solver->var_value(get_op_var(spec, i, j))) {
                             kitty::set_bit(op, j); 
                         }
                     }
@@ -307,7 +270,7 @@ namespace percy
 
                 // TODO: support multiple outputs
                 chain.set_output(0, 
-                        ((dag.get_nr_vertices() + spec.get_nr_in()) << 1) +
+                        ((spec.nr_steps + spec.get_nr_in()) << 1) +
                         ((spec.out_inv) & 1));
 
                 /*
@@ -319,7 +282,8 @@ namespace percy
                         continue;
                     }
                     for (int i = 0; i < spec.nr_steps; i++) {
-                        if (solver->var_value(get_out_var(spec, nontriv_count, i))) {
+                        if (solver_var_value(this->solver, 
+                                    get_out_var(spec, nontriv_count, i))) {
                             chain.set_output(h, ((i + spec.get_nr_in() + 1) << 1) +
                                     ((spec.out_inv >> h) & 1));
                             nontriv_count++;
@@ -330,9 +294,9 @@ namespace percy
                 */
             }
 
-            // Assumes that a solution was found by the synthesizer. In that
-            // case, we can block the current solution by blocking the current
-            // assignment to the operator variables.
+            /// Assumes that a solution was found by the synthesizer. In that
+            /// case, we can block the current solution by blocking the current
+            /// assignment to the operator variables.
             void
             block_solution()
             {